+++ oc.json
@@ -1,0 +1,9 @@
+{
+  "mocks": {
+    "plugins": {
+      "static": {
+        "getValue": "./value.js"
+      }
+    }
+  }
+}
